from netbsd; 64-bit cleanup

+++ sys/lib/libsa/exec.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: exec.c,v 1.10 1995/08/04 07:37:03 thorpej Exp $	*/
+/*	$NetBSD: exec.c,v 1.11 1995/10/20 00:47:47 cgd Exp $	*/
 
 /*-
  * Copyright (c) 1982, 1986, 1990, 1993
@@ -95,7 +95,7 @@ exec(path, loadaddr, howto)
 		goto shread;
 	addr += x.a_text;
 	if (N_GETMAGIC(x) == ZMAGIC || N_GETMAGIC(x) == NMAGIC)
-		while ((int)addr & (N_PAGSIZ(x) - 1))
+		while ((long)addr & (N_PAGSIZ(x) - 1))
 			*addr++ = 0;
 
         /* Data */
